backend/pyproject.toml
[tool.poetry]
name = "modurank"
version = "1.0.0"
description = ""
authors = ["junah201 <junah.dev@gmail.com>"]
license = "PRIVATE"
readme = "README.md"
include = ["lambdas", "shared"]
exclude = ["scripts"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {version = "^2.10.5", extras = ["email"]}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"
boto3-stubs = {version = "1.35.44", extras = ["dynamodb"]}
boto3 = "^1.36.2"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
